SharePoint Auth - Improved README and use the Token Cache (#21591)
After some fighting to get the SharePoint module to work, I want to
improve the SharePoint Authorization module with my findings.

1. Improved the README.md file
- Added information and examples of API Permissions and Scope needed to
get the authorization to work.
    - Corrected the syntaxes of the functions
2. Added support for Token Cache when acquiring an Access Token
- This removes unnecessary popups of the authorization window, and I
believe it is faster
- This makes it possible to schedule jobs that works against SharePoint,
as long as the user is authenticated now and then (I believe the default
life for a Refresh Token is 3 months)
3. Removed the custom caching of Access Token.
- "SharePoint Authorization Code" had implemented it's own caching of
the Access Token, with a hardcoded lifetime of 1 hour. From what I read,
that would probably work for most cases since the lifetime of an Access
Token is between 60 and 90 minutes. But as stated
[here](https://learn.microsoft.com/en-us/azure/active-directory/develop/active-directory-configurable-token-lifetimes),
the lifetime of an Access Token can be configured to a shorter time,
that would not work together with this "custom" token cache. A custom
caching must read the Access Token lifetime to be water proof, but the
easy fix here is to just remove it. :)
    - This change leaves the caching to the OAuth2 module.
4. Reworked the logic around the `IsSuccess` return variable.
- This had some issues, where `AcquireTokenByAuthorizationCode()` would
return true (it is a TryFunction, and some issues with acquiring an
access token does not result in an `Error()`), but with an empty Access
Token. This made me spend waaay to much time troubleshooting this
codeunit... The only thing that really tells us if acquiring an Access
Token failed or succeeded is if the Access Token has a value or not.

The tests are unaffected by this change.