Added time zone management module (#21246)
This module aims to provide methods to correctly calculate the offset
between a datetime in a time zone from UTC and the offset between time
zones. Other than the functions in the Type Helper codeunit in the base
app, this module takes daylight saving time into account when
applicable.

Additionally, the module provides indicators on whether a time zone
supports daylight saving time or whether a given datetime falls within
the daylight saving time period for a given time zone.