Dynamics GP Historical Transactions extension (#21440)
This is the new Dynamics GP Historical Transactions extension. It
includes tables to store historical GP transactions for G/L,
receivables, payables, purchase receivables, and inventory transactions.
It also contains pages to display this information.

- Settings added to allow for snapshot configuration.
- The GP Intelligent Cloud extension queues a background job at the end
of the migration to migrate the historical transactions snapshot.
- The GP Detail snapshot can be reran if needed.
- The Dynamics GP History SmartLists extension is deprecated and is no
longer installed when the migration is finished.
- Factboxes added to Customers and Vendors to display their historical
transactions.
- Factboxes added to the Cloud Migration Management page to show current
status, log, and errors.