Add OnBeforeSearchTable publisher for "Search in company data" feature (#25055)
**Issue Description**

The "Search in company data" feature lacks the possibility to filter the
results shown to the user.

There are some scenarios where displaying every single record that
contains a given **Search String** is not desirable.

While "security filters" are included in the "Data Search," they may be
insufficient if you have requirements to manage access to specific
tables on a record level for multiple users in various scenarios. In
some cases, this would require a consultant to create a separate
permission set for each user and thats obviously a lot of effort.

**Solution**

This PR introduces new ```OnBeforeSearchTable``` publisher, which allows
to prefilter the ```RecRef``` that is used for searching for the
```SearchString```.