use "if" "then" with get for posted document when calling e-document subscriptions (#25289)
In our BG localization we have a functionality based on Bulgarian
legislation that allows to void a posting document. Therefore, during
voiding, the header of an posted document is deleted, which is taken
with GET without checking in the e-document app. For now we remove
e-document for our customers to work with our localization. Please
change the code i think it will not cause errors.<br /><br />Internal
work item: AB#492530

Co-authored-by: Selime Halilova <s.halilova@TeamVISIONBulgaria.onmicrosoft.com>