GP - December 2023 enhancements (#25613)
This PR contains the following enhancements to the GP migration.

- Bug fix: When a company is deleted (usually after a failed migration),
delete from the GP Migration Error Overview table for the company. This
prevents an issue where subsequent migrations will be seen as failed
when they actually may not have been.
- Bug fix: Call ClearLastError() if the result of
MailManagement.CheckValidEmailAddress is false. This prevents future
attempts to obtain the last error text being that invalid email address
error text.
- Bug fix: Batch gets deleted even when the batch fails to post.
- Historical Snapshot smarter start logic. It will prefer to start as a
Background Job, but will fall back to a new Session if the user lacks
the necessary permissions to create a Background Job.
- Added missing Intelligent Cloud permission for table "GP CM20600".
- Rework to Item posting process. The posting logic has been moved to
where the other entity posting logic exists. Also, updated the batching
logic to actually have many transactions in a batch instead of just one.
This could be considered a performance improvement.
- Added additional information to checkbook/bank transactions
	- Document No. = CMTrxNum (currently from CMRECNUM)
	- Description = paidtorcvdfrom (currently DSCRIPTN)
	- External Document No. = CMLinkID (currently not set)
	
Fixes AB#496425

---------

Co-authored-by: jaymckinney <jay.mckinney82@gmail.com>