GP - Enhancements to Payment Terms creation, visibility of unposted batches, and Migration Log (#25831)
This PR enhances the GP migration by hardening the code around Payment
Terms creation, adds visibility of unposted batches, and a new Migration
Log.

### Payment Terms
This change hardens the code around Payment Term creation. It also
includes an integration event so that a failing Payment Term could be
created by a PTE if needed.

### Migration Log
This change also includes a migration log. A Factbox tile is added to
the Cloud Migration Management page that will show the user when a
non-error action was taken by the migration app, and they should be made
aware of it.

Currently, these are the scenarios that are now logged:

- The Payment Term was handled by 3rd party code.
- Skipped creating an Account because there isn't an account number.
- Skipped creating a PO because the Vendor has not been migrated.

### Unposted Batches
A Factbox tile is added to the Cloud Migration Management page that
shows the number of unposted batches.
Fixes
[AB#499261](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/499261)

---------

Co-authored-by: jaymckinney <jay.mckinney82@gmail.com>