[Shopify] Integration to use Sales Unit of Measure for inventory synchronization (#26707)
This pull request does not have a related issue as it's part of delivery
for development agreed directly with @AndreiPanko

Quick summary:

We saw that the Sales Unit of Measure should be used for inventory
synchronization but there is a bug that prevents it from doing it. In
line 61, Item.???Sales Unit Of Measure??? will always be empty, because
in the function in line 51, it is cleared. It means that if the
statement in line 63 never is true, the Base Unit of Measure will always
be used for inventory synchronization.

Fixes #26719
Fixes
[AB#539103](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/539103)

Co-authored-by: aidasberesinevicius <aidas.beresinevicius@1clickfactory.com>