[Shopify] Add Translations Export to Products and Variants (#26216)
This pull request does not have a related issue as it's part of delivery
for development agreed directly with @AndreiPanko

here‘s a quick summary:

Shopify Shop has a new action that takes the user to the Languages page.
Languages cannot be manually inserted or deleted, they can only be
refreshed with an action that pulls them from Shopify.
Before a language can be enabled for syncing, a language code needs to
be mapped to it, as Shopify is using locales, and I can‘t directly map
them.

A new Scope is required to get shop locales – read_locales

On export of products and variants, an update of translations is
triggered. The translation update first collects all
TranslatableContentDigests for a resource, as these are needed when
requesting an update of a translation.
The update next collects translations for the product Title, Product
Body, and Variant „option 1“ in a temporary table for each enabled
language. It compares the translations with the translations that were
last synced. If the translation is the same, it removes it from the
buffer. Next, the buffer is used to build the query and sent to Shopify
to update translations.




Continuing with a couple of questions/considerations regarding the issue

• The way I designed Shopify Languages is that the user can‘t
insert/delete them, but has to pull them from Shopify with an action.
This uses ShopLocales query which needs a read_locales scope that‘s not
currently part of the scopes. I wasn‘t sure if that‘s an issue or not.
• I made the import of locales skip the primary language as I wanted to
keep the logic for that one on the Shop.“Langauge Code“ field.
• I‘ve refactored the CreateProductBody procedure in the Product Export
codeunit so it takes the language code as a parameter. This means I can
reuse it for translations, however, there are events in this procedure
at the start/end, which will now trigger more often if more languages
are synced. I added the language code as a parameter in the publishers,
and technically nothing is broken until additional languages are synced,
but I wanted to mention this.
• I wasn‘t sure how I should approach translations when they‘re missing.
Should I take the existing approach, take the translated text or default
text, or should I only take the translated text, and if it doesn‘t
exist, skip that property?
This is mainly an issue with the product body. If Ext. Text is
translated, but Attributes are not, should I skip on attributes or send
attributes in the default language?
Currently, I skip the Title if it‘s not translated, but I send the body
in every case. Let me know if that should be changed.
• For the update of a translation, I need to send the
TranslatableContentDigest as part of the request. Currently, I‘ve done
this in a way, that I collect all digests for a resource during an
update, but that means that every update of a product/variant now has an
additional „get“ call, before it maybe sends an update for translations.
I think this could be done better, and would appreciate your input here.
• Resource Type is almost the same enum as used in Metafields, I think
these two should be made the same.
• I‘m not a fan of what I did with Translations Mgt. Is there a better
place where I could put this procedure for retrieving the item
translation?

Fixes #26819
Fixes
[AB#460580](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/460580)

---------

Co-authored-by: Tine Staric <tine.staric@companial.com>
Co-authored-by: Jesper Schulz-Wedde <JesperSchulz@users.noreply.github.com>