Salesperson API (#27664)
#### Summary
A new API page for Salespeople/Purchasers
I created the tests in a similar manner that I found suitable for
Locations, but I did not manage to run the test. Thus, looking forward
to the feedback if tests do pass in Microsoft Pipelines or how to test
those calls in a container. Currently I do not have a clue how to do it.

#### Work Item(s) 
Fixes #26806

Fixes
[AB#541735](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/541735)

---------

Co-authored-by: Jesper Schulz-Wedde <JesperSchulz@users.noreply.github.com>