Make dates language agnostic in Excel Reports (#28452)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to AlAppExtensions please read our pull request guideline
below
* https://github.com/microsoft/ALAppExtensions/blob/main/CONTRIBUTING.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Print sheets have had breaking formulas removed from dates and are split
so the date component is formatted as a Long Date:

![image](https://github.com/user-attachments/assets/969089ee-4445-473d-84fc-e77e91e28235)

About the report sheets include an additional time cell and UTC cell to
replace the datetime value lost on print pages.

![image](https://github.com/user-attachments/assets/6e5b0498-787c-4620-b1f0-073674f57b86)

That the ReportRequest.Date name is shown in UTC has been communicated
to Microsoft and a change is being discussed for conversion to local
time for clarity among users.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #28451

Fixes
[AB#573364](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/573364)

---------

Co-authored-by: Dan Williams <dan.williams@fenwick.com.au>