#28567 Update SubContractsGeneralMgt.Codeunit.al (#28582)
CheckIfVendorContractExistWhenDeleteVendor and
CheckIfCustomerContractOrServiceObjectExistWhenDeleteCustomer should
only be used if the master data record (vendor or customer) is not
temporary which is going to be deleted

#### Summary
Add a check to OnBeforeDeleteEvent event subscribers for both Customer
and Vendor tables to bypass contract validation when records are
temporary. This avoids unnecessary validation for temporary records,
which do not represent persisted data and should not trigger deletion
errors.

#### Work Item(s)

Fixes #28567 



Fixes
[AB#575819](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/575819)