Subscription Item and Default Item Quantity (#28800)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to AlAppExtensions please read our pull request guideline
below
* https://github.com/microsoft/ALAppExtensions/blob/main/CONTRIBUTING.md
-->
#### Summary <!-- Provide a general summary of your changes -->
In order to enable entering Subscription Items on Sales Orders and
posting of the Sales Order without an error few improvements or fixes
were done:
1. The function IsSalesLineWithSalesServiceCommitments has been improved
in the sense that it always returns a deterministic value for a sales
line regardless of the fact if the line has already been inserted and
thus "Line No." generated or not.
2. The function CalculateCalculationBaseAmount relies on up-to-date
sales line record. This was so far done by (a bit hacky) call of a
SalesLine.Modify on two places. Instead of modifying sales lines in the
middle of processing a new function has been introduced in "Sales
Subscription Line" table: SetSalesLine.
3. After fixing the code, the test
TestTransferSalesServiceCommitmentsOnExplodeBOM started failing altough
the manuall test was sucesfull. Therefore, the function
CreateComponentItemWithSalesServiceCommitments has been adjusted in
order to reflect the real use case..

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #28751


Fixes
[AB#581593](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/581593)